Fix a few warnings

+++ src/parsing_aux.c
@@ -5,7 +5,7 @@
 
 SmolRTSP_DeserializeResult SmolRTSP_parse(
     size_t max_data_size, size_t data_size, const void *restrict data, const char *restrict fmt,
-    uint_least8_t entities_to_parse, ...) {
+    uint32_t entities_to_parse, ...) {
     va_list args;
     va_start(args, entities_to_parse);
